fix() : 리뷰 없는 상품 노출시킬 때 에러 버그 고침

+++ src/main/java/com/example/instrumentshop/Goods/Entity/Review.java
@@ -42,5 +42,5 @@ public class Review {
     private String reviewDate;
 
     @Column(name = "review_score")
-    private double reviewScore;
+    private Float reviewScore;
 }
